chore(keychain-sdk): move RequestTracker to internal pkg
