+++ keychain-sdk/internal/tracker/tracker.go
@@ -1,34 +1,34 @@
-package keychain
+package tracker
 
 import (
 	"sync"
 )
 
-type RequestTracker struct {
+type T struct {
 	rw       sync.RWMutex
 	ingested map[uint64]struct{}
 }
 
-func NewRequestTracker() *RequestTracker {
-	return &RequestTracker{
+func New() *T {
+	return &T{
 		ingested: make(map[uint64]struct{}),
 	}
 }
 
-func (t *RequestTracker) IsNew(id uint64) bool {
+func (t *T) IsNew(id uint64) bool {
 	t.rw.RLock()
 	defer t.rw.RUnlock()
 	_, ok := t.ingested[id]
 	return !ok
 }
 
-func (t *RequestTracker) Ingested(id uint64) {
+func (t *T) Ingested(id uint64) {
 	t.rw.Lock()
 	defer t.rw.Unlock()
 	t.ingested[id] = struct{}{}
 }
 
-func (t *RequestTracker) Done(id uint64) {
+func (t *T) Done(id uint64) {
 	t.rw.Lock()
 	defer t.rw.Unlock()
 	delete(t.ingested, id)
